Move to an immediate-mode style update strategy.
On a state update, instead of re-rendering the component whose
state was updated, re-render starting at the root of the updated tree.
This involves more computation, but avoids the need to store component
parameters in memory. In a future commit, this will be levaraged to
allow non-'static component props. This code change also significantly
reduces code complexity in places by replacing the backing tree
data structure with a simple id to vnode map.