Loosen lifetime requirement on as_ref.
Previously, using as_ref required the DynRef produced to be valid as
long as the lifetime 'a of the producing DynBox. This is overly
restrictive, so instead now DynRef takes an additional lifetime
parameter 'r for the lifetime of its data so that a DynRef need not
live as long as the data producing it.