Allow referencing temporaries in builders.
Previously something like &String::from("") would cause a compiler error
in a component builder because that referenced a temporary value. This
was an issue because components were previously built and rendered over
several statements. This commit consolidates those steps into one
statement so that temporaries may be passed to builders just as they
may to functions.