Only accept function pointers in hooks.
Previously, we accepted arbitrary FnOnce functions as initializers
for state hooks. However, this seems to be unnecessary flexibility
and also required an extra generic parameter that often needed to be
inferred by the user. This removes that friction. If it turns out
initializers with access to outside state are useful,
explicit_generic_args_with_impl_trait will soon be stabilized, allowing
us to add FnOnce() implementing initializers back without a breaking
api change.