Execute callbacks by re-rendering.
In order to permit non-'static user-provided callbacks in a future
commit, enable executing callbacks by triggering a rerender, then
executing the user-provided function within update_component instead
of needing to store it.

This is enabled by a new update_vdom parameter that optionally provides
a native event and the component id it should be delivered to. If the
provided id is not of a native component, the event will not be
delivered and instead will be discarded. The update_component function
receives this event, if it exists, and can execute a callback with the
data. A native renderer can dispatch an event to be delivered via this
mechanism through a dispatcher provided in create_component. A renderer
may opt not to use this mechanism.

This mechanism assumes that only one event will ever be active at a
time, which is usually a reasonable assumption. Note, however, that this
is not always guarenteed to be the case, as on some platforms, users
can trigger events to synchronously run within event callbacks. This
implementation handles this case by scheduling the event to be processed
in a future tree re-rendering later in the event queue. It is up to the
renderer to ensure that any event state does not go stale; the
current_target case has been handled on web.