Update tracking system to use gen instead of bool.
Previously Tracked values stored an updated parameter determining
whether or not it had been updated since last render. However, this
isn't flexible enough: it doesn't enable users to create their own
user-defined fine-grained reactivity, as shown by the necessity of the
vec hook. Furthermore, it would likely cause issues with future hooks
like effect, which might not be called every render and thus miss the
render in which an input value is marked as updated. Note that there
internally already was a gen tracking system to enable the vec hook.
This commit extends its applicability to all tracked values.

Providing each tracked value with a gen enables the store hook for
user-customizable fine-grained reactivity, which is kept here and
replaces the vec hook.