Change default nested tracked behavior, keep old option.
Previously using nested tracked and updated calls was error-prone for
a user without knowing the distinction between location-dependent and
key-dependent updates. A user testing whether a tracked value at a given
index in an array changed since the last render would run into stale
UI state without an intuitive understanding of why, which is bad DX.

In this PR the default treatment of nested tracked and updated
invocations is that all nested calls contribute to the generation
tested in parent calls. This eliminates the need to avoid
location-dependent code at the cost of eliminating fine-grained
reactivity.

To restore fine-grained reactivity, the experimental primitives of
tracked_keyed and updated_keyed have been introduced, which use the
old behavior when containing nested calls. The documentation for this
behavior, however, remains poor. In the future either a clean way of
explaining this distinction is needed, the distinction needs to be
encoded more intuitively into library APIs, or a completely new approach
will be needed.