Implement new component call syntax.
The previous component call syntax relied on macro syntax, which was
confusing to understand conceptually while also preventing code
formatting. The new syntax keeps the requirement for components to start
with capital ASCII characters, but replaces macro invocations with
function calls with a starting parameter of self. The syntax of named
arguments was necessarily changed to fit into standard Rust grammar.
While the necessity of passing self to every component adds verbosity,
this was judged to be the least ugly way of disambiguating component
calls from enum variant constructors.