Implement bump allocation for component parameters.
Component builder constructors are now passed a reference to a bump allocator
to allow for bump allocating data throughout the prop setting phase.

Currently, this is used to bump allocate the children passed into components as
props. This could be extended to, for example, bump allocate String
representations for input values implementing Display.